requirements.txt
absl-py==0.13.0
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.2
charset-normalizer==2.0.4
clang==5.0
contourpy==1.1.0
cycler==0.10.0
flatbuffers==1.12
fonttools==4.54.1
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.40.0
h5py==3.1.0
idna==3.2
imageio==2.9.0
importlib-metadata==4.8.1
importlib-resources==6.4.5
keras-preprocessing==1.1.2
kiwisolver==1.3.2
markdown==3.3.4
matplotlib==3.6.0
networkx==2.6.2
nptyping==2.5.0
numpy==1.22.4
oauthlib==3.1.1
opencv-python==4.6.0.66
opt-einsum==3.3.0
pillow==8.3.2
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pynrrd==1.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
pywavelets==1.1.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.3
scipy==1.7.1
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
tifffile==2021.8.30
toml==0.10.2
typing-extensions==4.12.2
urllib3==1.26.6
werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
keras>=2.4.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0